Some merge request changes

+++ tik/lib/include/tik/KernelInterface.h
@@ -6,9 +6,9 @@ namespace TraceAtlas::tik
     class KernelInterface
     {
     public:
-        KernelInterface(int index, int block);
+        KernelInterface(int index, int64_t block);
         int Index;
-        int Block;
+        int64_t Block;
         bool operator<(const KernelInterface &b) const;
     };
 } // namespace TraceAtlas::tik
